requirements.txt
Flask>=2.3.3
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.7
WTForms>=3.1.2
Werkzeug>=3.0.1
requests>=2.31.0
python-dotenv>=1.0.1
matplotlib>=3.8.0
numpy>=1.26.4
Pillow>=10.2.0
textblob>=0.18.0
nltk>=3.8.1
apscheduler>=3.10.5
email-validator>=2.1.1
twilio>=9.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9